Refactor Dao creation with DaoFactory
Previously every dao constructed the connection string from individual
connection string parts and created the connection with DriverManager. The main
class contained copypasted methods for creating individual Daos in which the
stage was checked.

The connection string creation was moved to a helper method in the main class.
The connection creation was refactored to its own class ConnectionProvider.
The dao creation was refactored to a DaoFactory which takes a boolean
"isProduction" and the previously mentioned ConnectionProvider.