Make logging in work with tests
Previously, we checked if the app was running in production and conditionally
enabled the login filter. Now, the App class exposes a package-private method
`enableTestLogin()` which does two things.

1. It enables a test endpoint over at GET /test/login, which creates and signs
in with (sets the client's session) a test user
2. It changes the link at the login page from the Github auth URL to lead to
this test login endpoint.

This way, even while running acceptance tests, the app can redirect to the login
page. The test endpoint is currently enabled when a Cucumber @Given such as "new
bookmark is selected" is ran.